backend/requirements.txt
fastapi
uvicorn
tensorflow
opencv-python
numpy
pillow
tensorflow_hub
